requirements.txt
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
inflection==0.5.1
pyairtable==2.2.2
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.2.0
